chore: EditViewModel split into multiple files
